requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0

python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1